requirements.txt
langchain_community
youtube-transcript-api
langchain
langchain_chroma
langchain-google-genai  
python-dotenv
streamlit